Refactored checkerfacade
