+++ packages/core/src/checker/checker-factory.ts
@@ -6,7 +6,7 @@ import { coreTokens } from '../di/index.js';
 import { LoggingClientContext } from '../logging/logging-client-context.js';
 
 import { CheckerChildProcessProxy } from './checker-child-process-proxy.js';
-import { CheckerFacade } from './checker-facade2.js';
+import { CheckerFacade } from './checker-facade.js';
 import { CheckerRetryDecorator } from './checker-retry-decorator.js';
 
 createCheckerFactory.inject = tokens(commonTokens.options, coreTokens.loggingContext, coreTokens.pluginModulePaths, commonTokens.getLogger);
