+++ packages/core/src/checker/index.ts
@@ -1,3 +1,3 @@
 export * from './checker-factory.js';
 export * from './checker-resource.js';
-export * from './checker-facade2.js';
+export * from './checker-facade.js';
